Refactor tool conversion logic and improve code clarity across core and examples
